+++ Plugins/CrashRecovery/CHANGELOG.md
@@ -1,4 +1,7 @@
 # [CrashRecovery](https://1lighty.github.io/BetterDiscordStuff/?plugin=CrashRecovery "CrashRecovery") Changelog
+### 0.1.3
+- Fixed crash if XenoLib or ZeresPluginLib were missing
+
 ### 0.1.2
 - Fixed crash screen being blank if a plugin failed to stop properly.
 - Now closes all modals, including the special types that are not stored inside ModalStack.