+++ Plugins/MentionAliasesRedux/CHANGELOG.md
@@ -1,4 +1,7 @@
 # [MentionAliasesRedux](https://1lighty.github.io/BetterDiscordStuff/?plugin=MentionAliasesRedux "MentionAliasesRedux") Changelog
+### 2.0.5
+- Fixed crash if XenoLib or ZeresPluginLib were missing
+
 ### 2.0.4
 - Fixed tags in chat not showing but throwing an error instead in canary
 - Fixed tags not showing in new friends list